Crypto_Analysis_random_state/requirements.txt
# Kripto Para Analiz Sistemi için Gerekli Kütüphaneler
# =========================================

# Binance API entegrasyonu için
python-binance==1.0.19

# Veri analizi ve işleme için
pandas

# Excel dosya işlemleri için
openpyxl

# Gradient Boosting modeli için
xgboost

# Model değerlendirme ve veri bölme için
scikit-learn

# Bilimsel hesaplamalar için
numpy
